more work on the leveling system

+++ src/main/java/com/huskytacodile/alternacraft/data/DataSerializerRegistry.java
@@ -13,7 +13,7 @@ import java.util.EnumMap;
 public class DataSerializerRegistry {
     public static final DeferredRegister<EntityDataSerializer<?>> SERIALIZERS = DeferredRegister.create(ForgeRegistries.Keys.ENTITY_DATA_SERIALIZERS, Alternacraft.MOD_ID);
 
-    public static final RegistryObject<EntityDataSerializer<EnumMap<AlternaDinoEntity.DinoLevelCategory, Integer>>> DINO_LEVEL_CAT = SERIALIZERS.register("dino_level_cat",
-            () -> EntityDataSerializer.simple((writer, map) -> writer.writeMap(map, FriendlyByteBuf::writeEnum, FriendlyByteBuf::writeInt),
-                    (reader) -> new EnumMap<AlternaDinoEntity.DinoLevelCategory, Integer>(reader.readMap(buf -> buf.readEnum(AlternaDinoEntity.DinoLevelCategory.class), FriendlyByteBuf::readInt))));
+    public static final RegistryObject<EntityDataSerializer<EnumMap<AlternaDinoEntity.DinoLevelCategory, Double>>> DINO_LEVEL_CAT = SERIALIZERS.register("dino_level_cat",
+            () -> EntityDataSerializer.simple((writer, map) -> writer.writeMap(map, FriendlyByteBuf::writeEnum, FriendlyByteBuf::writeDouble),
+                    (reader) -> new EnumMap<AlternaDinoEntity.DinoLevelCategory, Double>(reader.readMap(buf -> buf.readEnum(AlternaDinoEntity.DinoLevelCategory.class), FriendlyByteBuf::readDouble))));
 }
